[iOS] 마지막 upnext 저장 구조 변경
