+++ MiniVibe/MiniVibe/MiniVibeApp.swift
@@ -9,83 +9,13 @@ import CoreData
 import SwiftUI
 import EventLogKit
 
-struct PersistenceController {
-    
-    static let shared = PersistenceController()
-    let container: NSPersistentContainer
-    let managedContext: NSManagedObjectContext
-    let entity: NSEntityDescription
-    
-    init() {
-        container = NSPersistentContainer(name: "User")
-        container.loadPersistentStores { (storeDescription, error) in
-            if let error = error as NSError? {
-                print("Unresolved error \(error), \(error.userInfo)")
-            }
-        }
-        managedContext = container.viewContext
-        entity = NSEntityDescription.entity(forEntityName: "LatestUpnext", in: managedContext) ?? NSEntityDescription()
-    }
-    
-    func delete() {
-        let fetchRequest: NSFetchRequest<NSFetchRequestResult> = LatestUpnext.fetchRequest()
-        let deleteRequest: NSBatchDeleteRequest = NSBatchDeleteRequest(fetchRequest: fetchRequest)
-        _ = try? managedContext.execute(deleteRequest)
-    }
-    
-    func fetchTracks() -> [TrackInfo] {
-        var fetchResult = [LatestUpnext]()
-        var defaultTracksData = [TrackInfoData]()
-        var defaultTracks = [TrackInfo]()
-        let fetchRequest: NSFetchRequest = LatestUpnext.fetchRequest()
-        do {
-            fetchResult = try managedContext.fetch(fetchRequest)
-            fetchResult.forEach { defaultTracksData.append($0.track) }
-            print("FETCH RESULT : \(fetchResult.count)")
-            print("DEFAULT TRACK INFO DATA : \(defaultTracksData.count)")
-        } catch let error as NSError {
-            print("Could not fetch. \(error), \(error.userInfo)")
-            return []
-        }
-        print("FETCH RESULT : \(fetchResult.count)")
-        print("DEFAULT TRACK INFO DATA : \(defaultTracksData.count)")
-        for track in defaultTracksData {
-            print(track.title)
-        }
-        defaultTracksData.forEach { data in
-            print(data.title)
-            let artist = TrackInfo.Artist(id: data.artist.id,
-                                          name: data.artist.name)
-            let album = TrackAlbum(id: data.album.id,
-                                   title: data.album.title,
-                                   imageUrl: data.album.imageUrl)
-            let track = TrackInfo(id: data.id,
-                                  title: data.title,
-                                  lyrics: data.lyrics,
-                                  albumId: data.albumId,
-                                  album: album,
-                                  artist: artist,
-                                  liked: data.liked)
-            defaultTracks.append(track)
-        }
-        print(defaultTracks)
-        return defaultTracks
-    }
-    
-    func tracksToViewModel(tracks: [TrackInfo]) -> [TrackViewModel] {
-        var viewModelArr = [TrackViewModel]()
-        tracks.forEach {
-            viewModelArr.append(TrackViewModel(track: $0, eventLogger: MiniVibeApp.eventLogger))
-        }
-        return viewModelArr
-    }
-}
-
 class AppDelegate: NSObject, UIApplicationDelegate {
+    var nowPlaying: NowPlaying!
+    
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey : Any]? = nil) -> Bool {
-        let defaultTracks: [TrackInfo] = PersistenceController.shared.fetchTracks()
-        let trackViewModel: [TrackViewModel] = PersistenceController.shared.tracksToViewModel(tracks: defaultTracks)
-        MiniVibeApp.nowPlaying.upNext = trackViewModel
+        let defaultTracks = nowPlaying.dataManager.fetch()
+        let trackViewModel = nowPlaying.dataManager.tracksToViewModel(tracks: defaultTracks)
+        nowPlaying.upNext = trackViewModel
         return true
     }
 }
@@ -97,7 +27,6 @@ struct MiniVibeApp: App {
     static let eventLogger = EventLogger(local: nil,
                                          server: nil,
                                          reachability: ReachablilityObserver(hostName: "www.google.com"))
-    static let nowPlaying = NowPlaying()
 
     let persistentContainer: NSPersistentContainer = {
         let container = NSPersistentContainer(name: "Event")
@@ -105,10 +34,16 @@ struct MiniVibeApp: App {
         return container
     }()
     
+    let nowPlaying = NowPlaying()
+    
+    init() {
+        appDelegate.nowPlaying = nowPlaying
+    }
+    
     var body: some Scene {
         WindowGroup {
             MainTab()
-                .environmentObject(MiniVibeApp.nowPlaying)
+                .environmentObject(nowPlaying)
                 .onReceive(
                     NotificationCenter.default.publisher(
                         for: UIApplication.didBecomeActiveNotification)) { _ in
@@ -117,10 +52,10 @@ struct MiniVibeApp: App {
                 .onReceive(
                     NotificationCenter.default.publisher(
                         for: UIApplication.willResignActiveNotification)) { _ in
-                    PersistenceController.shared.delete()
-                    let upNext = MiniVibeApp.nowPlaying.upNext
-                    let tracksToSave = viewModelToTracks(viewModels: upNext)
-                    saveTracks(tracks: tracksToSave)
+                    nowPlaying.dataManager.delete()
+                    let upNext = nowPlaying.upNext
+                    let tracksToSave = nowPlaying.dataManager.viewModelToTracks(viewModel: upNext)
+                    nowPlaying.dataManager.saveTracks(tracks: tracksToSave)
                     MiniVibeApp.eventLogger.send(Background(userId: 0))
                 }
                 .onReceive(
@@ -134,42 +69,5 @@ struct MiniVibeApp: App {
                     MiniVibeApp.eventLogger.send(Terminate(userId: 0))
                 }
         }
-    }
-    
-    
-    private func viewModelToTracks(viewModels: [TrackViewModel]) -> [TrackInfo] {
-        var tracks = [TrackInfo]()
-        viewModels.forEach { tracks.append($0.track) }
-        return tracks
-    }
-    
-    private func saveTracks(tracks: [TrackInfo]) {
-        tracks.forEach { track in
-            print(track.title)
-            let albumData = TrackAlbumData(id: track.album.id,
-                                       title: track.album.title,
-                                       imageUrl: track.album.imageUrl)
-            
-            let artistData = ArtistData(id: track.artist.id,
-                                   name: track.artist.name)
-            
-            let data = TrackInfoData(id: track.id,
-                                     title: track.title,
-                                     lyrics: track.lyrics,
-                                     albumId: track.albumId!,
-                                     album: albumData,
-                                     artist: artistData,
-                                     liked: track.id)
-            
-            let savedInfo = NSManagedObject(entity: PersistenceController.shared.entity,
-                                            insertInto: PersistenceController.shared.managedContext)
-            savedInfo.setValue(data, forKey: "track")
-            do {
-                try PersistenceController.shared.managedContext.save()
-            } catch let error as NSError {
-                print("Could not save. \(error), \(error.userInfo)")
-            }
-        }
-        
     }
 }
